Rename setup module to 'setups'
